A surprisingly large commit to fix tracking of the path with Control. I've implemented my own versions of the offScreen[XY]D? and screen[XY]D? methods now to make the tracer slightly more "pixel is not at little square" (qv), even if it still does the box-filter reconstruction under magnification...

+++ tracing/PathAndFillManager.java
@@ -1336,7 +1336,7 @@ public class PathAndFillManager extends DefaultHandler implements UniverseListen
 		}
 	}
 
-	synchronized PointInImage nearestJoinPointOnSelectedPaths( int x, int y, int z ) {
+	synchronized PointInImage nearestJoinPointOnSelectedPaths( double x, double y, double z ) {
 
 		PointInImage result = null;
